+++ src/main/java/com/codetaylor/mc/artisanworktables/modules/worktables/integration/gamestages/PluginGameStages.java
@@ -2,8 +2,8 @@ package com.codetaylor.mc.artisanworktables.modules.worktables.integration.games
 
 import com.codetaylor.mc.artisanworktables.modules.worktables.api.WorktableAPI;
 import com.codetaylor.mc.artisanworktables.modules.worktables.integration.jei.PluginJEI;
-import com.codetaylor.mc.artisanworktables.modules.worktables.recipe.IRecipeWorktable;
-import com.codetaylor.mc.artisanworktables.modules.worktables.recipe.RegistryRecipeWorktable;
+import com.codetaylor.mc.artisanworktables.modules.worktables.recipe.IRecipe;
+import com.codetaylor.mc.artisanworktables.modules.worktables.recipe.RegistryRecipe;
 import mezz.jei.api.recipe.IRecipeWrapper;
 import net.darkhax.gamestages.capabilities.PlayerDataHandler;
 import net.darkhax.gamestages.event.GameStageEvent;
@@ -65,12 +65,12 @@ public class PluginGameStages {
         .getUnlockedStages();
 
     for (String name : WorktableAPI.getWorktableNames()) {
-      RegistryRecipeWorktable registry = WorktableAPI.getWorktableRecipeRegistry(name);
+      RegistryRecipe registry = WorktableAPI.getWorktableRecipeRegistry(name);
 
       if (registry != null) {
-        List<IRecipeWorktable> recipeList = registry.getRecipeList(new ArrayList<>());
+        List<IRecipe> recipeList = registry.getRecipeList(new ArrayList<>());
 
-        for (IRecipeWorktable recipe : recipeList) {
+        for (IRecipe recipe : recipeList) {
           String uid = PluginJEI.createUID(name, recipe.getTier());
           IRecipeWrapper recipeWrapper = PluginJEI.RECIPE_REGISTRY.getRecipeWrapper(recipe, uid);