+++ src/main/java/com/codetaylor/mc/artisanworktables/modules/worktables/tile/spi/TileEntityTypedBase.java
@@ -46,5 +46,4 @@ public abstract class TileEntityTypedBase
     return this.type;
   }
 
-  protected abstract EnumTier getTier();
 }