+++ src/main/java/com/codetaylor/mc/artisanworktables/modules/worktables/tile/spi/TileEntityWorkstationBase.java
@@ -61,7 +61,7 @@ public abstract class TileEntityWorkstationBase
   }
 
   @Override
-  protected EnumTier getTier() {
+  public EnumTier getTier() {
 
     return EnumTier.WORKSTATION;
   }